requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.1
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
markdown2==2.4.12
MarkupSafe==2.1.4
numpy==1.26.3
openai==1.10.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
sniffio==1.3.0
SQLAlchemy==2.0.25
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
